Improvements in counting module, refactoring which moved parse_num function into queries/__init__.py so it can be used in any qmodule
